removed baseComponentFunction
