+++ lib/actions/createEvent.js
@@ -4,7 +4,6 @@ const messages = elasticio.messages;
 const _ = require('lodash');
 
 const ApiClient = require('../apiClient');
-const baseComponentFunction = require('../baseComponentFunction');
 const processEvent = require('../processEventDataHelper').processEvent;
 
 function getCalendars(cfg, cb) {
@@ -42,11 +41,24 @@ function processAction(msg, cfg) {
     self.emit('data', messages.newMessageWithBody(messageBody));
   }
 
+  function emitError(e) {
+    console.log('Oops! Error occurred');
+    self.emit('error', e);
+  }
+  function emitEnd() {
+    console.log('Finished execution!');
+    self.emit('end');
+  }
+
   console.log("Creating Event with properties: ");
   console.log(postRequestBody);
 
   const apiCall = `/me/calendars/${cfg.calendarId}/events`;
-  return baseComponentFunction(self, apiCall, postRequestBody, emitData, cfg, 'Calendars.ReadWrite');
+  const instance = new ApiClient(cfg, 'Calendars.ReadWrite');
+  instance.post(apiCall, postRequestBody)
+    .then(emitData)
+    .fail(emitError)
+    .done(emitEnd);
 
 }
